SDC-466 - Create deployment infra for spark
Creates SparkManager class which encapsulates interacting with the spark cluster
and adds hidden script .spark-manager script which is used to actually interact with
the cluster.

The object ApplicationState wraps the YARN App ID and can be used to get status.

Stag libraries specified in the pipline configuration will automatically be shipped.

Change-Id: Ia08e61727e002dcc0d77fd82b4c8080799033c39
Reviewed-on: https://review.streamsets.net/144
Reviewed-by: Brock Noland <brock@streamsets.com>