SDC-1316 - SDCClassLoader does not correctly canocialize META-INF resources and calls super when delegating to parent
Both xerces and org.xml/wc3 must be loaded by the same cl and org.xml/w3c is provided by the jvm.

Given this we have two bugs in our cl setup:

1) Some if not most JREs request META-INF resources without the starting / which we don't correctly handle.
this was causing xerces version of DocumentBuilder to be loaded out of the stage cl when it should
have been delegated to the parent.

2) The getResources method when delegating to the parent calls super as opposed to parent. This results
in the fix for bug #1 to not actually apply.

Change-Id: If7d873a120b73aa38069744b955a0785772b7676
Reviewed-on: https://review.streamsets.net/838
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Brock Noland <brock@streamsets.com>