SDC-3302: Introduce MathEL covering java.lang.Math methods
I've added a new EL class that covers most useful methods from
java.lang.Math class. I've skipped all the arc/cos and other geometric
functions as they did not seem particularly useful to us. We can add
them in future JIRAs if/when needed.

I also wanted to honor the initial type, so for example when one calls
math:abs(int) the output will be int and similarly math:abs(double) will
return double (and the same for float and long). That slightly
complicates the code but it seems as better option then clobber the
types and causing unnecessary drift in the data.

I've introduced one allowed casting and that is from string to double or
long. I'm assuming that this will be required because some of sources of
data are inherently String based (record headers, CSV files). The
current logic is fairly dump - if the string contains dot, it's
assumed to be double, otherwise long. Any parsing error will result in
EL exception.

Change-Id: I16421ca6d5326db5118540c8a6b0208fe4b1e6e0
Reviewed-on: https://review.streamsets.net/3307
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>