SDC-3137: Test: Add non-default database to ColdStartIT
I've refactored the ColdStartIT a bit to also run against both "default"
and "custom" databases to make sure that we will create the tables
properly.

I've also added a simple ParametrizedUtil class that will generate cross
product of various independent arguments for us (so that we don't have
to iterate those combinations manually).

Change-Id: I26ebeb00d2e066bdf0124261aaf22994baa05a0c
Reviewed-on: https://review.streamsets.net/3109
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>