SDC-3222. Add Header attributes support for JDBC column extra information.
We need some extra information for the columns. The easy way to do this populate some headers attributes.
To make sure user is aware of this and not override, all headers will be prefixed with a jdbcNsPrefix (default - "jdbc.")
For now the main use of this to populate scale and precision for decimal columns (as BigDecimal's scale and precision is not much dependable.)
This is currently will be used by the default scale and default precision EL configurations at HiveMetadataProcessor.
Introduced a new EL for getting the current field being processed by the processor, this is just a context variable for now.
The FieldPath EL is used only in HiveMetadataProcessor currently.

Change-Id: I504e55864135576741bf04a950eea00950f9fcdf
Reviewed-on: https://review.streamsets.net/3201
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>