SDC-3196. Compose slave SDC Id using master SDC Id and partition id
Currently slave SDC id is a random UUID. Now we are changing it to be more
meaningful and composing using master SDC id and a unique id (In case of
Spark this will be the partitionId and in case of MR it will be the task id)
The advantage is now even if a worker restarts to work on a partition/task, it will
get the same id as before. This will be useful as we can preserve the partition/task metrics
across worker restarts.

Change-Id: I64c74c9bd4946cec32fac5ca96cdd06ff0e3d0db
Reviewed-on: https://review.streamsets.net/3242
Reviewed-by: Virag Kothari <virag@streamsets.com>