SDC-3246: Test: Enhance AllPartitionTypesIT
Added all our types to the test to make sure that they are failing in
expected manner. I had to do changes to the main code by adding an extra
validation because, we're "not supporting" some types for partition by
not allowing them in the UI. I wasn't sure how to test the UI, so I've
added a hard check instead. I believe that the cost in init() is
negligible, so we should be fine.

I've skipped testing Decimal as that will lead to ArrayOutOfBounds
exception as precision and scale isn't added - and we currently don't
even have a way how to specify those in the partition configuration, so
I guess we should be good there as well.

Change-Id: I472f314a65043f98acf11c8e98c7bbd7678c539a
Reviewed-on: https://review.streamsets.net/3223
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>