SDC-3257: Refactore JDBC Origin to explicitly cast types based on metadata
Currently the JDBC Origin is casting types more implicitly based on
class returned by ResultSet.getObject() that is transformed to our type
in JsonUtil.jsonToField(). This however breaks in several scenarios:

* All NULL constants are converted to STRING regardless of what type the
  column is. This is generating unnecessary drift that is causing issues
  for example in Hive Metadata processor.
* All three time based types - DATE, TIME and TIMESTAMP are mapped to
  our type DATE because JDBC will return Date() object for them. This is
  however very bad as for TIME and TIMESTAMP we are loosing information.

Instead of using the JsonUtil.jsonToField() method, I've provided
explicit method that is transferring JDBC types to our field types. This
way we have full control on type mappings.

I've also added code to serialize the java.sql.Types constant of each
column to the Record header if createJDBCNsHeaders is set to true. This
way users have the ability to know what was the original type in DB and
act accordingly if needed.

Change-Id: I8400040df4c8e0bbbf303c7717265f0d0c837811
Reviewed-on: https://review.streamsets.net/3236
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>