SDC-3259: Change default value for getting decimal precision and scale to not use any defaults
After seeing real deployments, I came to conclusion that the default value
we have in Decimal's scale and precision is more hurtful then helpful. If
the there are no headers containing the scale/precision value, then we
default to 2 and later in the code throw:

  HIVE_26 - Invalid decimal value 1023609: precision 7 is more then
  expected 2

That is however quite confusing as user is expecting that the number
might be wrong where in fact only the configuration is incorrect. Hence
I've changed the default configuration for scale/precision to not
contain default value and force users always specify valid expression.

I've also changed the EL expression to expect a string and we're doing
our own conversion to int. This has to be done because the EL library
will automatically and silently convert any invalid string number to "0"
rather then throwing exception.

Change-Id: Ia00c8199b32664cf0472c0fceffaf5aebadfe873
Reviewed-on: https://review.streamsets.net/3238
Reviewed-by: Junko Urata <junko@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>