SDC-3256. Handle location mismatch for partitions. This fix basically handles the case where the partition values are same but the partition location is changed, we will route the corresponding record to error record, so user can take a corrective action in hive metadata processor.
In HMS Target as well we will route the record to error,
But the files are written to hive in both locations.
in that case hms target will error out, user can pick and choose what location
they want later by manually executing the command to add partition.
This mostly applies to the external table.

Change-Id: I6a7615d01b24a42b3eb2f641549074c3de1395cc
Reviewed-on: https://review.streamsets.net/3247
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>