SDC-3270: Remove UTF-8 characters from TestBinaryDataParser and TestBinaryDataParserFactory
It seems that one had copy pasted string from a website when creating
TestBinaryDataParser and TestBinaryDataParserFactory containing a
non-ascii characters. Then both tests are calling getBytes without
specifying Charset and hence are using the platform default. This is
causing troubles where the platform default is different then the test
author intended. I would suggest to simply substitute the extra UTF-8
characters with ASCII ones, so that it works regardless of OS/env
configuration.

Change-Id: I71445f8d99602099c552e5f1d9f2fe6c75f6f4fe
Reviewed-on: https://review.streamsets.net/3253
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>