SDC-3308. HDFS Target commit writers resets idle closed flag
Once a writer is idle closed, the HDFS target commits the writer, which in turn calls RecordWriter#close() which resets the idleClosed flag to false.
In the close() method in the writer, we now check if it is idle closed and throw, which is caught in the commit writers anyway and logged.

Change-Id: I1fb062b3d215cf5d62b6733c33154facf4ef0747
Reviewed-on: https://review.streamsets.net/3323
Reviewed-by: Adam Kunicki <adam@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>