SDC-3350: Time: Add new type to CsvUtil, JsonUtil and XmlUtil
Added the TIME to all three util classes. Two of them seems to be unused
whereas the third one (JsonUtil) seems to be used here and there. They
do not currently have a good test coverage and I don't want to introduce
proper test coverage as part of this patch.

Change-Id: I38e6a3cb7a681bc9d8eae4ab94b57a5a2670c46b
Reviewed-on: https://review.streamsets.net/3365
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>