SDC-3110. File Tail Origin throws StringIndexOutOfBoundsException when calculating Pending Files JMX Metric.
For Log Files with reverse counter (For EX: file tail process .log.2, .log.1 etc...) when all files with reverse
counter are processed, the live file will be one without counter, in this case we should report
pending files to 0, but the comparator does not understand files without counter.
This fix adds a check to see if the current file is an acceptable file if not return 0.

Change-Id: Ic5a1ef2af476f7c84a29461d1cb91e10e0af6168
Reviewed-on: https://review.streamsets.net/3454
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 5ce719385517402b256de50159d44dfb4bcb2e80)
Reviewed-on: https://review.streamsets.net/3472
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>