SDC-3239. Add MapR support for Hive metadata processor, Hive Metastore Target and Hive Target.
This fix adds support for running multiple instance of hive target (streaming target). There
was a Library loader which should have been loaded by only one class loader but given that
HiveStreaming says privateClassLoader=true, multiple instance of hive streaming running in a
single sdc was throwing Linkage Error
(instance of  sun/misc/Launcher$ExtClassLoader): attempted  duplicate
class definition for name: "com/mapr/fs/shim/LibraryLoader"

The fix was to make sure it is loaded by only the api class loader similar to
how we added Jni classes of MAPR

The same error happened to hive drift solution pipeline. This is because there are two stages in the pipeline
using mapr hive -> hive metadata processor and hive metastore target. Apart from that mapr hive-exec jar
which is needed for connection with hive packages avro 1.7.3 (this is just in MAPR) and streamsets uses 1.7.7.
There is a setValidateDefaults in Schema$Parser class which is available only in 1.7.7
and causing NoSuchMethodException in MAPR stages, this fix disables usage of the API.
There is a impact because of this, if user specified wrong default values over AvroTypeUtil
will not validate

This fix also fixes the hdfs location regex for partition to add maprfs:/ as MAPR locations
will contain maprfs:/ prefix

Change-Id: Ie4deb102d0b1e3822c6c7b49730df49907572481
Reviewed-on: https://review.streamsets.net/3429
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>