SDC-3401: Explicitly validate that query contains WHERE and ORDER clauses
In JDBC origin we require the query to contain both WHERE and ORDER BY
clauses currently. We also have check that they are in proper order and
that the offset column is first in both clauses. However when user
misses one of the clauses the error exception is quite confusing Query
must include '{}' in WHERE clause and in ORDER BY clause before other
columns.. I would like to propose improving the exception and issue a
direct validation error Hey you're missing X clause.

Change-Id: Ida75033ca6fc0f19d74d705c7a066c219b1191d4
Reviewed-on: https://review.streamsets.net/3448
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>