SDC-3457. Add Whole File Format.
Introduces Whole File Data Format along with Data parser and Data Generator.

Whole File Data Parser:
  The Whole file Data Parser creates a record for each file with a FileRef
Field and a metadata field for each file, so each record parsed is one for
single file

Whole file Data Generator:
   The Whole File Data Generator uses the record created by the Whole File Data
Parser which contains information about the file and the File Stream Ref for
creating a stream for reading, the generator treats one record as a whole file
and transfer it in one shot (internally buffering and writing).

Any errors during this operation will have to restarted from scratch, this
indicates file transfer will not have a feature of resumability.

Error handling using different generator format:
	Most of the existing data generator formats (TEXT, BINARY, DELIMITED)
use getValueAs function which will by default fail during fileRef conversion.
For Avro, Protobuf there is an augumented which will make sure the conversion
fails.

Metrics:
 	The target will produce two metrics to keep track of the file transfer
progress.
	1) File Tranfer rate (bytes/sec)
	2) Remaining bytes to be transferred.

This patch also edits the existing sources to add this as a dummy group
as the data format value is not in the list for supported data format
chooser values.

Upgrade:
   Given that all are new configurations introduced with default values,
there is not an upgrade impact.

Checksum Verification:
        The target when it reads the file also calculates the checksum and
compare it with the checksum (if provided) in the file metadata of the record
If the source system does not support checksum verification the record will be
an error record.
If the source system automatically verifies the checksum on file open the
verifyChecksum configuration is ignored.
If the source system provides the checksum in the file metadata and does
not automatically verify the checksum, the target when it reads the file
compares thc checksum during the stream close call.

This patch also edits the existing sources to add this as a dummy group
as the data format value is not in the list for supported data format
chooser values.

Change-Id: I2b667f76f93fd80c4b01893183297134a58f9e41
Reviewed-on: https://review.streamsets.net/3722
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>