SDC-3481. HDFS Target idleExecutor should set removeOnCancel to true.
Since each RecordWriter actually creates a task per batch to close itself on idle,
we should make sure that the executor removes old tasks, which it does not, by
default causing a lot of memory to be used up for no real purpose.

Unfortunately, this method is not exposed by the `ScheduledExecutorService`,
so we use the actual implementation, `ScheduledThreadPoolExecutor`.

Additionally, we only adjust the idle future only on batch flush now, rather than
individual record writes.

Change-Id: Ie28eaeb52bb2f49a36c67e33a01f6f4457f20c46
Reviewed-on: https://review.streamsets.net/3631
Reviewed-by: Adam Kunicki <adam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>