SDC-3535: Enhance File Tail Origin to produce events
The origin currently produces events on a data line (that is called
"metadata"). This patch also publishes the same information on event
lane - e.g. this change is fully backward compatible. The idea is to
provide users an upgrade path (e.g. few releases when they will be
able to switch from the data to event lane) and at some point in the
future when we release a "breaking" release, then we will drop
sending the events on the original data lane.

Change-Id: I3d01bc2fe46f9269727078193d07e9e02a62f279
Reviewed-on: https://review.streamsets.net/3740
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>