SDC-3533: Introduce events and event lanes to pipes
Introduced an even lane concept to the Pipe. Only StagePipe is allowed to
have even lane, all other pipes have only input and output lanes.

Since the Pipes are a runtime concept (they are not persisted anywhere)
I've decided to not introduce inputEventLane and outputEventLane methods to
all Pipes and rather work with the Event lanes (outside of StagePipe) in
the same way as with any other output lane.

Change-Id: Iad53c70bb32d9c92cbbd0dfd4f4b6f72787a8b36
Reviewed-on: https://review.streamsets.net/3729
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>