SDC-3572. Add Support of Whole file in S3Source.
This patch adds support for whole file format in s3Source.
A verify checksum checkbox for verifying checksum when the
data is read from s3 which will be compared with
metadata header in s3.

The current limitation (which applies for Directory Source)
is that each batch produces only one record which is the
metadata along with fileref object for whole file.

Change-Id: I75774861785e9d5774195cc4cc2cd6b752aabdef
Reviewed-on: https://review.streamsets.net/3899
Reviewed-by: Harikiran Nayak <hari@streamsets.com>