SDC-3588. Change the date format for avro schema serialization of hive metastore to use UUID.
As of now if the hive metastore configuration does not specify useAsAvro, Avro schema is serialized in hdfs(whenever schema changes/table created) and
hive metastore is updated to set the latest location in tblproperties. Inorder to make sure the schema files are not overwritten by one another
(especially the stage is multiplexed for multiple tables), the avro schema is serialized in avro_schema_${db}${table}${date as yyyy-MM-dd--HH-mm-ss}.avsc,
but this format is problematic as the date is (only till seconds) and two schema change record on the same table in the same second can cause it to use
the same path and throw Alread schema exists IOException.

To avoid the above condition this patch will replace the data substring in the file name with UUID, so the resulting filename will be
avro_schema_${db}_${table}_${uuid}.avsc

Change-Id: I592cbe89def9e2cc7ec5836b4033b4671f998967
Reviewed-on: https://review.streamsets.net/3776
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>