SDC-3617: DPM in CM: Create base code to register SDC in DPM
I've added several new DPM-related properties to our descriptor file for
Cloudera Manager. As part of those properties user needs to specify if
DPM should be enabled and if so, it's URL and credentials to a role that
have ability to register SDC. Then on next restart of the service, we
use DPM REST calls to pre-generate the app token and store it outside of
CM space (so that it can be reused on service restart) and configure SDC
to use this token.

The whole process is thus driven from CM and user don't have to do any
manual intervention outside of filling the CM configs and restarting the
service. This patch doesn't contain ability to re-generate the token, that will
be a follow up work.

Change-Id: I83007ca8e69dedd7f46638b1bf5a87eeb6af602d
Reviewed-on: https://review.streamsets.net/3819
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>