SDC-3594: Create check box 'Skip verifying directory' in HDFS/Local FS destination
To describe the use case a bit more. Our current HDFS init call will
always try to create some directory (or file) to verify that we have
sufficient permissions to HDFS. That however doesn't work very well for
the following case:

* Directory template is /a/b/${record:attribute('c')}/d/e. Nobody can
  write into /a/b, all the ${record:attribute('c')} are pre-created.

* On init we don't have any record and thus ${record:attribute('c')}
  evaluates to empty string which in turn will evaluate to path /a/b/d. As
  that doesn't exists we will try to create it. Since nobody can write into
  /a/b/ and we end up with an exception during init.

The root cause is that we're evaluating the path without any record
which might lead different path on init that during normal execution. One
way how to workaround it is with attributeOrDefault, however that still
requires some valid value of the default directory name.

Change-Id: I7fdde82f354fecbc8e44f9a91cdcd9f40949edaf
Reviewed-on: https://review.streamsets.net/3821
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>