SDC-3538: Add support for producing and consuming events while in destroy phase
This patch tweaks semantics of the destroy action a bit - rather then
destroying each pipe one by one after all have been finished, the
destroy operation is now delegated to PipelineRunner and it's expected
that all pipes after the current one are still "alive" while destroying.

This enables us to iterate through the list of pipes and one-by-one do
last process() of the pipe and then immediately destroy it. As a
optimization all stages on data path are immediately destroyed without
the last process() call as we know that there will be no data.

Change-Id: If79dade58c422d5644c81cfdf6fce43f410f40b9
Reviewed-on: https://review.streamsets.net/3849
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>