SDC-3638. HDFS Target caused deadlock when idle close is enabled.
The deadlock happens when PipelineRunnable executes
flushAll and idleCloseThread executes writers.release() in RecordWriter.close()

This is because the pipeline runnable calls flushAll which is a
synchronized method (takes a lock on ActiveRecordWriters) and
tries to take a read lock on writer by calling writer.isClosed()

At the same time, idle close thread calls close(true) -> which
calls writers.release(writer, roll) -> which first takes a
writer.closeLock() which takes a write lock on writer and
then tries to remove it from the writers by
  synchronized(this) {
	 writers.remove(writer.getPath().toString());
  }

The synchronized block tries to take a lock on ActiveRecordWriters
but cannot as the pipeline runnable called flushAll and already has
it.
Similarly the pipeline runnable cannot take a lock on the writer in
writer.isClosed() has the idle thread already took a write lock.

To avoid this whenever both locks are needed at the same time,
the ordering is first take lock on ActiveRecordWriters and then
the writer.

In this case the IdleCloseThread will first take a lock on ActiveRecordWriters
and then call close()
Also changed the release method to be synchronized.

Change-Id: I5ae6997042ab9567b1684c86a77ee0edb9847ab0
Reviewed-on: https://review.streamsets.net/3862
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>