SDC-3139. Force-quit a pipeline
Normal stop pipleine senario is, when the stop button is pressed,
the main thread gets interrupted and exits from loop (run() in ProductionPipelineRunner),
then pipeline status is changed from STOPPING to STOPPED.

However the issue is that when script processsor has infinite loop, the main thread
that runs script engine ignores interrupt and keeps running, so that the main thread
is stuck in loop and pipleine status can never be changed.

User has two options:
We had a bug in #1 case. After waiting for 5 minutes, we were not changing the pipeline state
to STOPPED. Added fix for that.

However in both situation, the main thread that runs script engine still keeps running.
It looks like no way to terminate the thread unless we call thread.stop().
Would love to have solution for this.

Change-Id: I3bc56d34fc85f4714307b028ad47d3a8f9cceef8
Reviewed-on: https://review.streamsets.net/3967
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>