SDC-3711: Hive metadata processor is calculating partition value and path using different time
We're currently not sharing the ELVars with calculated timeDriver when
we calculate the partition value and partition path (for external
table). This is a problem because both calculations are using different
time - the difference is usually below millisecond, but it could happen
that this millisecond will cause us calculating different value for each
field.

This patch changes the behavior to always use the same ELVars with the
same calculated timeDriver for all time-driven calculations for one
record.

Change-Id: I30cbb3faf4abda46b606102326b946afebfa26bd
Reviewed-on: https://review.streamsets.net/4007
Reviewed-by: Junko Urata <junko@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>