SDC-3648: NPE when trying to tail a file and permission is denied
I've just ensured that there will be NPE thrown to pollute the logs. All
the previous functionality remained the same. I'm not sure how to easily
reproduce the issue in unit tests, so this patch does not came with one.

It's worth pointing out that the current behavior of the code is to skip
the files it can't load and this behavior had remain the same. I'm not
entirely sure whether i's the best one or not, but that figuring it out
is a task for separate JIRA.

Change-Id: I4f1639fc80ed5e276cf800d0f67bb63c87e1a6ab
Reviewed-on: https://review.streamsets.net/4053
Reviewed-by: Adam Kunicki <adam@streamsets.com>