SDC-3551. Directory origin may skip processing some files when they all arrive at the same time.
If a file comes in at the same millisecond as the listFiles call,
and we have already seen a file which is lexicographically greater than the new file
with the same timestamp (which is also the timestamp as the listFiles), we end up
ignoring the new file.

The fix is pretty simple, we just pick up files whose timestamp is strictly less than
the timestamp of the current scan.

Change-Id: Ieeb4f67a530e81f434971160b6d028fb8c52a4b3
Reviewed-on: https://review.streamsets.net/4112
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>