SDC-3496. Directory origin archiving with timestamp order throws exceptions
When the current file has already been archived (due to it being processed),
the Files.getLastModifiedTime method call in the pathComparator.compare ends up
throwing a NoSuchFileException.

This patch simply catches this exception and then adds the file anyway (since if it
was not consumed, then it means it is newer).

Change-Id: I3fc92c546c05deb69a8949c6d361c0f55f62e6eb
Reviewed-on: https://review.streamsets.net/4139
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>