SDC-3775. Move CDH5.7 and Apache Kudu 0.7 and 0.8 inside all-libs profile in dist project pom.
The general process is to move the old versions to all-libs profile in dist project (which packages the product)
After CDH 5.8 is added and apache kudu 0.9 is added, the old versions are still left behind in default profile.

This patch moves the dependencies of CDH 5.7 and Apache Kudu 0.7 and 0.8 libs from default profile
to the all-libs profile.

Change-Id: I5c9ad1d162c5586f1fe83dd6b53a5a5b7f0aa716
Reviewed-on: https://review.streamsets.net/4152
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>