SDC-3301. File Tail Origin - 'WARN Error while Calculating Offset Lag' when File Post Processing - Delete
The patch address the follwing issue, the normal flow of things is:
1. getOffsets for files
2. Read files
3. setOffset
4. calculateOffsetLag and Pending files for each file information configured

The calculation for offset lag is done by using the filenames in the offset using (Files.size(filename) - currentOffset)
and it may so happen that the file is archived/deleted after the file is fully processed in that case, we are getting NoSuchFileException

The patch basically checks whether the fileContext has a currentReader and then try to see whether it is done with the file (reader == null).
If thats the case, skip calculating the offset lag.

Change-Id: I554ed89f530e18731c684cc7a6b8a172abbef512
Reviewed-on: https://review.streamsets.net/4215
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>