SDC-3576.  Value Replacer should allow substitution of specific values
Enhanced Value Replacer to permit replacing Byte, Short, Integer,
Long, Float, Double and String datatypes with new values based on the
original values.

There is a new section in the UI "Conditionally Replace Values" which
contains lists into which the user can enter field names, the operator
(<, =, >, ALL) and the value to compare against. There is also a field
for the value which will be placed into the field when the condition is
met.

Bigdecimal, DATE, TIME, and DATETIME datatypes are not supported.

Less than, equals, and greater than all work with the field's native data
type. for example, in the case of an Integer, the numbers in the
"Comparison Value" and "Replacement Value" are converted to
Integer for comparisons.

When selecting "ALL" as the operator, the "Replacement Value" is converted
to the appropriate datatype and placed into the field unconditionally.

As with other features of Value Replacer, multiple fields can be added
to the "Fields To Replace" entry field.

The conditional replace is last to run, first "Fields To Null" runs, then
"Replace Null Values".  After both of those functions complete,
"Condionally Replace Values" runs.

Processing is top to bottom, so a field which is specified multiple times,
will be processed multiple times in order.

For example:
field to replace  | operator | comparison value | replacement value
department             =        100                   50
department             =        200                   300

This will process a field multiple times and may overlay a previous
replacement...

Example 2
field to replace  | operator | comparison value | replacement value
department             =        100                   50
department             =        50                    200
this will yield 200 as a result for departments which were originally 100.

Updated tests, too.

Change-Id: I34c72806f80c3868d6285046325234d7ef654b78
Reviewed-on: https://review.streamsets.net/4234
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>