SDC-3835.  Field Converter should be able to convert a byte array to a string
Modified Field Converter to support decoding byte [] to String.  Also
added a new feature to the UI permit the user to select the destination
String encoding.

The default encoding is UTF8.  In some cases the user will have to know the
encoding of the data in the byte [] to get proper results.  The byte []
does not have a notion of the encoding of the data it contains.

Change-Id: I8c1f34b55338a6d31efc31bfc8110c5e388809c5
Reviewed-on: https://review.streamsets.net/4260
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>