SDC-3840. SDC doesn't update the Kinesis checkpointer correctly when the records from multiple shards error out
If kinesis checkpointer don't sync up with seqno (last offset), it's possible to throw an invalid seqno in different shardID.
This can happen in certain scenarios, for example if we have one valid record stored in shard1 and one invalid record stored in shard2 with maxBatchSize with 3. If commit was called with valid record seqno and checkpointer of shard2, then validation check will fail (since seqno is in shard1).
This patch is fixing this issue to make sure seqno always sync up with corresponding checkpointer.

* Customer saw the error saying "with certain scenario, Kinesis checkpointer and Sequence Number
Change-Id: I9b69b7d3feb335395dfc6d18eb42475193c49839
Reviewed-on: https://review.streamsets.net/4265
Reviewed-by: Adam Kunicki <adam@streamsets.com>