SDC-3859: Update icons for the new stages
I've used icon of a package with arrow pointing out of the package to
represent the field flattener processor as I wasn't able to came up with
a better one. For the other new components, I've decided to continue
using the current icons as I don't have a better one for the executors.

Change-Id: I669f3b2f76d058cbbe8aad096772f78ad8d85d6a
Reviewed-on: https://review.streamsets.net/4273
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>