SDC-3760: Build sporadically fails to generate static web directory
Prior to this patch, there was no dependency declared between ui and
dist module so for example on my machine, maven decided to build dist
before ui. This resulted in dist using ui build in previous run as we're
directly accessing files in the ui module. This patch adds this
dependency as provided to force maven build them in the right order.

Change-Id: I72de72747918eb667c2eb766b32e0817f9b72bde
Reviewed-on: https://review.streamsets.net/4274
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>