SDC-3900. NPE when using SFTP if the file is processed in multiple batches.
This patch address the issues when NPE is thrown if the file is processed in
multiple batches.
The issue started happening after the remoteFile argument is added to the function
addRecordsToBatch which is set to null at the start of produce function eventhough
parser is not null, the remoteFile is null and this caused an NPE because the new
header attributes tries to get the file name from the remoteFile object.

Change-Id: Ib134355a58fb97954de456a13be48eb996bb6615
Reviewed-on: https://review.streamsets.net/4316
Reviewed-by: Bob Plotts <bob@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>