SDC-3001.  Enhance SFTP/FTP with directory traversal and file selection.
Added 2 new fields to the UI in the SFTP/FTP tab.

"Process Subdirectories" is a checkbox and added a text field for
"File Name Selection".  When "Process Subdirectories" is checked,
all the subdirectories of the specified directory are traversed, searching
for files to process.

"File Name Selector" supports a limited file glob comprised of * and ? along
with literals.  Tilde, dot and slash are NOT permitted at the beginning of
the glob, dot is treated as a literal elsewhere in the glob.  The glob in
this field limits the files which are selected such as ??a for
three-character file names which end with 'a'.  *.txt represents all files
with file names of one or more characters ending with ".txt".

"File Name Selection" works in conjunction with "Process Subdirectories"
checkbox - so traversal can also have the file mask applied.
Directory names are not checked - only file names are checked.

Change-Id: Ic0602b87ae947a185089b345b680fd8b86add3b1
Reviewed-on: https://review.streamsets.net/4356
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>