SDC-3951. AWS Whole File Format throws Resetting to invalid mark error on Retry.
This patch addresses the error retries causing an invalid mark on reset during file transfer.
Given that SDC uses offset mechanism, instead of S3 client retrying SDC will retry based on ATLEAST_ONCE/AT_MOST_ONCE semantics.

The issue relates to the fact the mark position is no longer valid in SdkBufferedInputStream in S3Client during the reset (when an error happens, for ex: network glitch and there will be a subsequent retry which tends to try a reset on the stream). Most of the seekable streams can do this properly but given that we won't be able to mandate all streams from different source are seekable, we will make sure there won't be a retry from the s3 client (so no reset).

The semantics are the whole file will be retried by the SDC.

Some related links for the issue:
1.https://www.instaclustr.com/blog/2016/04/20/cassandra-backups-using-aws-transfermanager/
2.https://github.com/aws/aws-sdk-java/issues/427

Change-Id: I9e466b3fb2d75d8908123dca2238115f1475dc23
Reviewed-on: https://review.streamsets.net/4431
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Cheolsoo Park <cheolsoo@streamsets.com>