SDC-3989: HDFS config file validation prevents HDFS Origin from functioning
The problem was that YARN uses symlinks quite a lot in the temporary
directories - the resources directory ends up being behind at least on
symlink. But since we were running the getCanonicalPath() only on the
calculated hadoop conf-dir path and not on the resources dir, we got
the exception.

This patch fixes the problem by running the same getCanonicalPath() on
the resources directory as well to get consistent point of view.

Change-Id: I126444e91cac5da25f8f05877dd0417263e5fef7
Reviewed-on: https://review.streamsets.net/4472
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>