SDC-4005: Introduce RecoverableDataParserException in data parser library
The basic idea is that there are two types of parser exceptions:

* Unrecoverable. Corruption is so hard that we’re not even able to tell
where one record ends and second starts and once this exception is
encountered, rest of the stream (file) is not usable.

* Recoverable. Corruption is such that we can’t construct one record, but
know where it ends and where another (and probably proper one) starts.
Example would be a line in CSV file that have more column separators
then it should (as the separators in data are not properly escaped).

We currently throw the same exception for both cases, so I propose to
create a new subclass of DataParseException called
RecoverableDataParseException that will be thrown by the shared library
if and only if it occurred an error during parsing that does not invalidate
the entire stream. This exception will require an accompanying Record object
containing all information that human being needs to re-create the malformed
input such that it can be reprocessed after fixing the malformation.

Since the RecoverableDataParseException is a child of DataParseException all
currently existing stages will work the same way without change and it’s up
to the stage implementation to opt-in into processing the new exception
differently - and even possibly expose the behavior as a configuration to
the end user.

Change-Id: Ie6a4e86414e80911f3da9f2c12cd6da77da98aaa
Reviewed-on: https://review.streamsets.net/4478
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>