SDC-4033: Common-el have global cache for all ELs and static values
As described in EL-1 the class ExpressionEvaluatorImpl have two global
caches - sCachedExpressionStrings and sCachedExpectedTypes. All executed
ELs - and for some ELs also values are persisted in those caches and
never cleaned up.

This patch provides custom subclass of the ExpressionEvaluatorImpl that
replaces the implementation from Map to LRUMap to ensure that they will
not grow out of bound on us.

Change-Id: I12c05c3604f466be21c63df1f40cd0f458a5f11b
Reviewed-on: https://review.streamsets.net/4560
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>