SDC-3860: Check number of file descriptors at startup
Rather then going with Java solution as implied on the ticket, I've
decided to make a shell solution that won't even start the java project
unless the limit for file handlers is large enough.

I've also written the check in a way that user can easily set the
expected value to 0 and thus completely disable the check if he so
chooses. The default value is set to 4k even though that we currently
recommend 8k to not break any devs - the default value on mac is 4864.
Also I know that unless user is doing something really crazy, the 4k
actually should be sufficient. We can increase it later to something
higher if needed.

Change-Id: I84b5e0dab09cce1afb7e2811c7970bb9a4f216da
Reviewed-on: https://review.streamsets.net/4576
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>