SDC-4057. HDFS Shutdown hook may cause SDC Shutdown hook to not end gracefully on hdfs stages
Whenever SIGTERM is issued on a standalone sdc or stop on a cluster pipeline is issued (which internally kills the job), SDC calls the shutdown hook which makes sure the pipeline is gracefully shutdown.
During this shutdown hook the destroy is called on each stage which performs some hdfs operations (for ex: in hdfs target -> we rename the files from _tmp)

Shutdown hooks are thread, so if there are multiple shutdown hooks registered the order is not guaranteed
It seems like DFsClient in HDFS also has a shutdown hook defined in
private final ClientFinalizer clientFinalizer = new ClientFinalizer(); in FileSystem$Cache
which runs in an shutdown hook closing all the cached filesystem objects.

This for example will fail the rename operation we will try to do in hdfs target. This patch sets fs.automatic.close = false in hadoop configuration which will disable the shutdown hook of hdfs to close all cached file system objects.HDFS Stage destroy() will close the filesystem object which was created.

Change-Id: I33680558c337b6fb29b8a0983e2ba1f07bda1365
Reviewed-on: https://review.streamsets.net/4609
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>