SDC-3650 - EL for converting Unix time to human readable date/timestamp
Added several routines to the EL time library:

millisecondsToDateTime — takes 1 arg, the epoch time in milliseconds,
and returns a Java Date object.

extractStringFromDate -  takes2 args, first
is the Date object; the second is the date format specifier.

extractLongFromDate - takes 2 args, the first is a Date object;
the second is the date format specifier. If overflow or another
error occurs (such as an invalid format specifier)
an exception may be thrown.

extractStringFromDate and extractLongFromDate are particularly
well suited for creating database partitioning keys.
The expression:
${time:extractLongFromDate(record:value('/theDate')), "MM")}

will return only the month as a long value.  Adding this to a
new column in the record can provide the partitioning key.

Note: since these routines return a Long - the non-numeric data items
are not going to work in this format for example (a = AM/PM indicator,
G = Era designator); using MMM will return the 3-character abbreviation
for the month (such as Jul), and will cause the method to throw an
exception or return incorrect results.

Date format specifier details can be found here:
https://docs.oracle.com/javase/7/docs/api/java/text/SimpleDateFormat.html

Change-Id: I42d71c8e7fd32fb6f663ed003150497f2ac5f10f
Reviewed-on: https://review.streamsets.net/4632
Reviewed-by: Bob Plotts <bob@streamsets.com>