SDC-2363. Directory origin should support subdirectory traversal
Added support for traversing an arbitrarily deep directory
structure starting from the directory specified in the Files Directory field.
This new feature only works when the files are being processed in
chronological order - this feature is not enabled when processing
file in lexicographical order.

A new checkbox was added - Process Subdirectories which is
visible only when the Read Order listbox is set to Last Modified Timestamp.

Files are selected based on the glob or regex in File Name Pattern.
Files are processed in last modified timestamp order regardless
of their location in the directory tree.

When using the new Process Subdirectories option, and using the
Post Processing features, such as Error processing or Archiving, the
partial path names are preserved and the file will be moved to
the corresponding position in the Error or Archive directory structure.

If, for example, the source Files Directory is set to /data/logs and the
Error Directory is set to /data/errors when /data/logs/sub/thefile.log is
processed (and has an error), it is moved to /data/errors/sub/thefile.log -
keeping the original subdirectory structure in place.

Change-Id: I47099cd1b7339d520229f03706c8fed727706a1a
Reviewed-on: https://review.streamsets.net/4636
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>