SDC-4178. REST call /v1/pipelines/status should ignore metrics property
We store pipeline metrics data in Pipeline Status JSON when pipeline is 
stopped, so that we can display pipeline metrics for previous run in 
history tab. 

It is used only in pipeline page > history tab. 

In home page we don't use 'metrics' property at all. So when we call 
REST API /v1/pipelines/status in home page we should ignore metrics 
property to reduce the pay load.

Change-Id: Ib4bfb922283209457fcf33e4b60e8a920b92857d
Reviewed-on: https://review.streamsets.net/4795
Reviewed-by: Harikiran Nayak <hari@streamsets.com>