SDC-4224. Remove limit on the number of oldest SCNs selected in Oracle CDC
In some Oracle environments, V$ARCHIVED_LOG may contain rows corresponding to
log files that have already been removed. In this case there may be more than
10 logs that don't exist. Since we limit the number of log files we look at to
find the oldest SCN, we might not be able to start LogMiner if all the oldest
10 logs that V$ARCHIVED_LOG corresponds to have already been removed. So we should
remove this limit and find the oldest one whichever it is.

Change-Id: I1a989f077037756c4ab76a34bbb4aabf1d10c8a2
Reviewed-on: https://review.streamsets.net/4857
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>