SDC-3970.  FieldRemover matches prefix of field.
Modified Field Remover to handle several situations where regexes were
not correctly processed.  If the user has 3 fields -
field1, field11, and field13 and requested matching field1 only, the logic
would match all 3 fields although only field1 was supposed to be selected.

Fixed a problem with * metacharacter within a word or at the end of a
field path.  Added support for ? as a meta character to represent
a single character anywhere in a field path.

Change-Id: Ice8c9b469e16f056e5dbe73fa118beb1afee5077
Reviewed-on: https://review.streamsets.net/4874
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>