SDC-3603.  Improve performance of FieldRenamer
Added a “cache” to FieldRenamer to hold the results
of the previous record’s schema processing.   The cache
is only one item; if the new record’s schema is different
in any way the cache is not valid, and the code runs
the old way.  The speedup is greatly configuration-dependent,
and is most visible when there are many columns to rename
in each record.

Also cached results from an expensive call - which may
make a difference whan the code runs "old way'.

The record’s schemas may not match if there is an upstream
“remove field if null” processor - perhaps reordering the
pipeline can fix this issue.  The cache is also not valid
if the incoming record has an array where there are
different numbers of elements in each record.

Change-Id: I3076f2c059a920df7af71318dd12e3f5e072c011
Reviewed-on: https://review.streamsets.net/4882
Reviewed-by: Adam Kunicki <adam@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>