SDC-4240: Update text/default for memory.monitor SDC property
Looking into history, the memory monitoring was implemented a long time
ago, but enabled by default across the board in 1.4.0.0 release by
SDC-1674. I suggest to turn it off by default for all customers since we
know that it's expensive and our new suggestion is to not use it for
production loads.

Change-Id: I7ad44c2ab367a6ae41f3148686ca613fa739afd2
Reviewed-on: https://review.streamsets.net/4907
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Loretta Chen <loretta@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>